Add a 'OneArgNew' trait
A variation on the theme that does away with giving
the type and init arg attribute name in
the `with`, but rather lifts it straight from the
attribute's declaration.

The additional code doesn't change anything at all (and doesn't interact)
with the role, so this could be released as an independent
MooseX::OneArgNew::AsTrait. If you prefer me to do that, just say the word.
I just thought that offering the two options in the same distribution
wouldn't be the worst thing ever. :-)